+++ src/main/java/fi/dy/masa/servux/Servux.java
@@ -3,6 +3,8 @@ package fi.dy.masa.servux;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import fi.dy.masa.servux.dataproviders.DataProviderManager;
+import fi.dy.masa.servux.dataproviders.EntitiesDataProvider;
+import fi.dy.masa.servux.dataproviders.LitematicsDataProvider;
 import fi.dy.masa.servux.dataproviders.StructureDataProvider;
 import fi.dy.masa.servux.event.PlayerHandler;
 import fi.dy.masa.servux.event.PlayerListener;
@@ -16,6 +18,8 @@ public class Servux
     public static void onInitialize()
     {
         DataProviderManager.INSTANCE.registerDataProvider(StructureDataProvider.INSTANCE);
+        DataProviderManager.INSTANCE.registerDataProvider(LitematicsDataProvider.INSTANCE);
+        DataProviderManager.INSTANCE.registerDataProvider(EntitiesDataProvider.INSTANCE);
 
         ServerListener serverListener = new ServerListener();
         ServerHandler.getInstance().registerServerHandler(serverListener);